requirements.txt
discord
python-dotenv
PyNaCl
yt-dlp
asyncio